Get Landsat SR Products

This function gets 30-m USGS Landsat 8, 7 , 5 and 4 (Surface Reflectance Tier 1 (SR)) satellite images from various bands. The image processing for time series can either use methods by Roy et al. (2016) and Zhu et al. (2015) (e.g. harmonization and cloud-masking) so that all three Landsat missions could be used for time series analysis ('harm_ts') or you can just get time series without harmonizing ('ts').

get_landsat(
  aoi,
  method = "ld8",
  param = NULL,
  stat = "median",
  cloud_mask = TRUE,
  startDate = "1984-01-01",
  endDate = "2020-10-30",
  mask = FALSE,
  m.low = NULL,
  m.high = NULL,
  c.low = 1,
  c.high = 12
)

Arguments

aoi

A sf object indicating the extent of the geom.

method

A character indicating what method to use, e.g. 'ld8', 'ld7', 'ld5', 'ld4', 'ts', 'harm_ts'.

param

A character indicating what band to visualize, e.g. 'Blue' or c('Green', 'Red', 'NIR') or nothing returns all bands.

stat

A character indicating what to reduce the imageCollection by, e.g. 'median' (default), 'mean', 'max', 'min', 'sum', 'stdDev', 'first'.

cloud_mask

logical whether to mask out certain cloud artifacts. TRUE (default).

startDate

character format date, e.g. "1999-10-23"

endDate

character format date, e.g. "1999-10-23"

mask

logical whether to mask out certain ranges

m.low

numeric low value for mask, e.g. greater than 'm.low'

m.high

numeric high value for mask, e.g. less than 'm.high'

c.low

numeric lower month value for calendar range

c.high

numeric higher month value for calendar range

Value

A list of Earth Engine Objects and arguments.

Details

The methods currently available (more to come):

harm_ts: Harmonizing Landsat Missions; 1984-01-01 - 2021-01-22.

ts: Combining Landsat Missions; 1984-01-01 - 2021-01-22.

ld8: LANDSAT/LC08/C01/T1_SR; 2013-04-11 - 2021-01-22.

ld7: LANDSAT/LC07/C01/T1_SR; 1999-01-01 - 2021-01-21.

ld5: LANDSAT/LC05/C01/T1_SR; 1984-01-01 - 2012-05-05.

ld4: LANDSAT/LC04/C01/T1_SR; 1982-08-22 - 1993-12-14.

The param (bands) currently available (more to come):

Blue, Green, Red, NIR,SWIR1, SWIR2, NDVI, NDWI, NBR

Note

If you want to use this with #eemont functions then make sure cloud_mask is set to FALSE. Also, 'harm_ts' and 'ts' can take a long time to process in other function viz, rr and band so be aware!

Examples

if (FALSE) {
# Load Libraries

library(rgee)
rgee::ee_intialize()
library(exploreRGEE)

# Bring in data
huc <- exploreRGEE::huc

ld8 <- get_landsat(huc, method = 'ld8', startDate = '2014-01-01',
                  endDate = '2018-12-31', c.low = 6, c.high = 11)

}