Get Meteorological Products

This function gets you all the Meteorological data you want in a Earth Engine object. See details for available methods.

get_met(
  aoi,
  method = "Norm81m",
  param = NULL,
  stat = "median",
  startDate = "1981-01-01",
  endDate = "2010-01-01",
  mask = FALSE,
  m.low = NULL,
  m.high = NULL,
  c.low = 1,
  c.high = 12
)

Arguments

aoi
A sf object indicating the extent of the geom.

method
character. 'Norm81m'(default), see details.

param
character. NULL (default). Use the band names for appropriate dataset method, e.g. PRISM = 'ppt', GRIDMET = 'pr', DAYMET = 'prcp', etc.

stat
A character indicating what to reduce the imageCollection by, e.g. 'median' (default), 'mean', 'max', 'min', 'sum', 'stdDev', 'first'.

startDate
character format date, e.g. "1999-10-23"

endDate
character format date, e.g. "1999-10-23"

mask
logical whether to mask out certain ranges

m.low
numeric low value for mask, e.g. greater than 'm.low'

m.high
numeric high value for mask, e.g. less than 'm.high'

c.low
numeric lower month value for calendar range

c.high
numeric higher month value for calendar range

Value

A list of Earth Engine Objects and arguments.

Details

The methods currently available (more to come):

Norm81m: PRISM 1981-2010; Monthly, annual average.

AN81m: PRISM Jan 1895 - ongoing; Monthly annual time series.

AN81d: PRISM Jan 1 1981 - ongoing; Daily time series.

GRIDMET: 1979-01-01 - ongoing; The Gridded Surface Meteorological dataset provides high spatial resolution (~4-km) daily surface fields of temperature, precipitation, winds, humidity and radiation across the contiguous United States from 1979.

DAYMET: 1980-01-01 - 2019-12-31; Daymet V3 provides gridded estimates of daily weather parameters for United States, Mexico, Canada, Hawaii, and Puerto Rico. It is derived from selected meteorological station data and various supporting data sources.

TRMMh: 1998-01-01 - 2020-12-31; The Tropical Rainfall Measuring Mission (TRMM) is a joint mission between NASA and the Japan Aerospace Exploration Agency (JAXA) designed to monitor and study tropical rainfall. The 34B2 product contains a gridded, TRMM-adjusted, merged infrared precipitation (mm/hr) and RMS precipitation-error estimate, with a 3-hour temporal resolution and a 0.25 degree spatial resolution.

TRMMm: 1998-01-01 - 2019-12-01; This collection is no longer being updated. See IMERG monthly. This dataset algorithmically merges microwave data from multiple satellites, including SSMI, SSMIS, MHS, AMSU-B and AMSR-E, each inter-calibrated to the TRMM Combined Instrument.

TERRACLIMATE: 1958-01-01 - 2019-12-01; TerraClimate is a dataset of monthly climate and climatic water balance for global terrestrial surfaces.

Examples

if (FALSE) {
# Load Libraries

library(rgee)
rgee::ee_intialize()
library(exploreRGEE)

# Bring in data
huc <- exploreRGEE::huc

prism_monthly <- get_met(huc, method = 'AN81m', startDate = '2014-01-01',
                  endDate = '2018-12-31', c.low = 6, c.high = 11)

}